upload
There is a value in the register $s0. Write a piece of MIPS code that puts in $s7 that value multiplied by 31. There are the following conditions:

    You can use only add, addi, sub instructions.
    Your code should have not more than 10 instructions.

Breaking any of these conditions leads to 0 points 

Hint: If you have a number X and you add it to itself, you receive X+X=2X. So, adding the number to itself is the same as multiplying it by 2.